WIP: Destroy Associations in Background Job Motivation:   - Sometimes cascading association deletions can cause timeouts due to   IO issue. Perhaps a model has associations that are destroyed on   deletion which in turn trigger other deletions and this can continue   down a complex tree. Along this tree you may also hit other IO   operations. Such deep deletions can lead to server timeouts while   awaiting completion and really the user may not notice all the   changes on their side immediately making them wait unnecesarially or   worse causing a timeout during the operation.
  It is possible to move the association deletion into jobs
  today, however it is pretty boilerplate and it would be nice to make
  it easier to do.

  This pr builds off the work of @georgeclaghorn in PR #34901 and is
  currently a simple proposal of how this could work. I branched off his
  pr for now to quickly product this proof of concept.

Related Issues:
  - #34901
  - rails/actionmailbox#16

Changes:
  - Added a new job to handle deleting associations.
  - Added a new `dependent` option on `has_many` associations,
  `destroy_async` (Open to other names).
  - Added a single proof of concept test.

+++ activerecord/lib/active_record/associations/has_many_association.rb
@@ -26,6 +26,9 @@ module ActiveRecord
           # No point in executing the counter update since we're going to destroy the parent anyway
           load_target.each { |t| t.destroyed_by_association = reflection }
           destroy_all
+        when :destroy_async
+          load_target.each { |t| t.destroyed_by_association = reflection }
+          enqueue_destroy_all
         else
           delete_all
         end
